requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.4.4
SQLAlchemy==1.3.24
Werkzeug==2.0.3
PyMySQL==1.0.2
cryptography==3.4.7
python-dotenv==0.19.0
pytest==6.2.5 